Experimental sharding support
This is not being merged into master yet as I'm still not sure of the complete effects of sharding the bot out. Plus, it isn't super necessary at this time.

When the bot starts getting a little larger, I'll merge this, but for now, it'll remain in the testing phase.